Update en-replace.pl.pairs (#4)
* Update en-replace.pl.pairs

It's a first draft. I changed some obvious mistakes. It's not that easy to know the meaning without a context in Polish language. For example:  `Medium` could be translated into `Średni` if you are taking about medium height or `Średnia` if you are taking about price because price is feminine and height would be neutral in Polish language.